Increase a bit the acces token lifespan
The default value is 5 minutes which is a pain when using the API via
Swagger for example.
The recommendation is for this value to be lower than the SSO idle
timeout which is 30 minutes by default.

We should probably configure all the timeout properly, but for now 30min
on the access token should make for a better dev experience